+++ src/components/Shared/Jumbotron.tsx
@@ -12,7 +12,7 @@ import {
 } from '@material-ui/core'
 import { Link } from 'gatsby'
 import Particles from 'react-particles-js'
-import ocLogo from '../../assets/images/four51-white.svg'
+import ocLogo from '../../assets/images/four51-logo--white.svg'
 
 const styles = (theme: Theme) =>
   createStyles({
